+++ IBS.DataAccess/Migrations/20260708075808_RemoveNotificationFeature.cs
@@ -1,0 +1,93 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace IBS.DataAccess.Migrations
+{
+    /// <inheritdoc />
+    public partial class RemoveNotificationFeature : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "hub_connections");
+
+            migrationBuilder.DropTable(
+                name: "user_notifications");
+
+            migrationBuilder.DropTable(
+                name: "notifications");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "hub_connections",
+                columns: table => new
+                {
+                    id = table.Column<Guid>(type: "uuid", nullable: false),
+                    connection_id = table.Column<string>(type: "text", nullable: false),
+                    user_name = table.Column<string>(type: "text", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("pk_hub_connections", x => x.id);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "notifications",
+                columns: table => new
+                {
+                    notification_id = table.Column<Guid>(type: "uuid", nullable: false),
+                    created_date = table.Column<DateTime>(type: "timestamp without time zone", nullable: false),
+                    message = table.Column<string>(type: "text", nullable: false),
+                    target_url = table.Column<string>(type: "text", nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("pk_notifications", x => x.notification_id);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "user_notifications",
+                columns: table => new
+                {
+                    user_notification_id = table.Column<Guid>(type: "uuid", nullable: false),
+                    notification_id = table.Column<Guid>(type: "uuid", nullable: false),
+                    user_id = table.Column<string>(type: "text", nullable: false),
+                    is_archived = table.Column<bool>(type: "boolean", nullable: false),
+                    is_read = table.Column<bool>(type: "boolean", nullable: false),
+                    requires_response = table.Column<bool>(type: "boolean", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("pk_user_notifications", x => x.user_notification_id);
+                    table.ForeignKey(
+                        name: "fk_user_notifications_application_user_user_id",
+                        column: x => x.user_id,
+                        principalTable: "AspNetUsers",
+                        principalColumn: "id",
+                        onDelete: ReferentialAction.Cascade);
+                    table.ForeignKey(
+                        name: "fk_user_notifications_notifications_notification_id",
+                        column: x => x.notification_id,
+                        principalTable: "notifications",
+                        principalColumn: "notification_id",
+                        onDelete: ReferentialAction.Cascade);
+                });
+
+            migrationBuilder.CreateIndex(
+                name: "ix_user_notifications_notification_id",
+                table: "user_notifications",
+                column: "notification_id");
+
+            migrationBuilder.CreateIndex(
+                name: "ix_user_notifications_user_id",
+                table: "user_notifications",
+                column: "user_id");
+        }
+    }
+}