+++ IBS.DataAccess/Repository/IRepository/IHubConnectionRepository.cs
@@ -1,9 +1,0 @@
-namespace IBS.DataAccess.Repository.IRepository
-{
-    public interface IHubConnectionRepository
-    {
-        Task SaveConnectionAsync(string username, string connectionId);
-        Task RemoveConnectionAsync(string connectionId);
-        Task RemoveConnectionsByUsernameAsync(string username);
-    }
-}